Fix trace, debug color and make sub-loggers chainable
